setup/requirements.txt
configparser
mysql-connector-python

